+++ firebase-perf/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
+
+# 20.5.1
 * [changed] Make Fireperf generate its own session Id.
 
 # 20.5.0